requirements.txt
python-docx
fpdf
PyYAML
